+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+## TBD
+
+### Enhancements
+
+* Span parentage APIs now require a BugsnagPerformanceSpanContext, which BugsnagPerformanceSpan is now a subclass of. You no longer need to assign a Bugsnag span as a parent. [280](https://github.com/bugsnag/bugsnag-cocoa-performance/pull/280)
+
 ## 1.6.1 (2024-06-24)
 
 ### Bug fixes